Simulation kernel : Clean up & warning fix

+++ src/libraries/simulation_kernel/Entities/Agents/Roles/Urban/PHY_RoleInterface_UrbanLocation.h
@@ -48,7 +48,7 @@ public:
     virtual double ComputeUrbanProtection( const PHY_DotationCategory& dotationCategory ) const = 0;
     virtual MT_Vector2D GetFirerPosition( MIL_Agent_ABC& target ) const = 0;
     virtual MT_Vector2D GetTargetPosition( MIL_Agent_ABC& target ) const = 0;
-    virtual float ComputeDistanceInsideSameUrbanBlock( MIL_Agent_ABC& target  ) const = 0;
+    virtual double ComputeDistanceInsideSameUrbanBlock( MIL_Agent_ABC& target  ) const = 0;
     virtual float ComputeRatioPionInside( const MT_Ellipse& attritionSurface ) const = 0;
     virtual float ComputeRatioPionInside( const geometry::Polygon2f& polygon, float modificator ) const = 0;
     //@}
